Updated code with review comments
