+++ src/LCT.Common/Constants/Dataconstant.cs
@@ -80,7 +80,6 @@ namespace LCT.Common.Constants
         public const string ModerationRequestMessage = "Moderation request is created";
         public const string Cdx_SpdxFileName = "internal:siemens:clearing:spdx-file-name";
         public const string SpdxImport = "SPDXImport";
-        public const string UnsupportedPackageType = "UnsupportedPackageIdentifiedFromSpdx";
         public static Dictionary<string, string> PurlCheck()
         {
             return purlids;
