+++ src/LCT.PackageIdentifier/Model/BomKpiData.cs
@@ -66,7 +66,7 @@ namespace LCT.PackageIdentifier.Model
 
         [DisplayName(@"Components Updated From SBOM Template")]
         public int ComponentsUpdatedFromSBOMTemplateFile { get; set; }
-        [DisplayName(@"Total SPDX components were imported as baseline entries")]
+        [DisplayName(@"Total SPDX components imported as baseline entries")]
         public int UnsupportedComponentsFromSpdxFile { get; set; }
 
         public string ProjectSummaryLink { get; set; }
